+++ ui/src/shared/parsing/flux/response.ts
@@ -59,10 +59,6 @@ export const parseTables = (responseChunk: string): FluxTable[] => {
   const tableColIndex = headerRow.findIndex(h => h === 'table')
   const timeColIndex = headerRow.findIndex(h => h === '_time')
 
-  if (!timeColIndex) {
-    throw new Error('Could not find time Column')
-  }
-
   // Group rows by their table id
   const tablesData: Array<Array<Array<string | number>>> = Object.values(
     _.groupBy(nonAnnotationData.slice(1), row => row[tableColIndex])
@@ -70,7 +66,9 @@ export const parseTables = (responseChunk: string): FluxTable[] => {
 
   const groupRow = annotationData.find(row => row[0] === '#group')
   const defaultsRow = annotationData.find(row => row[0] === '#default')
-  const dataTypeRow = annotationData.find(row => row[0] === '#datatype')
+  const dataTypeRow: string[] = annotationData.find(
+    row => row[0] === '#datatype'
+  )
 
   // groupRow = ['#group', 'false', 'true', 'true', 'false']
   const groupKeyIndices = groupRow.reduce((acc, value, i) => {
@@ -100,8 +98,23 @@ export const parseTables = (responseChunk: string): FluxTable[] => {
       {}
     )
 
+    const numberColumns = (dataTypeRow || []).reduce((acc, val, i) => {
+      if (val === 'long' || val === 'unsignedLong' || val === 'double') {
+        acc.push(i)
+      }
+      return acc
+    }, [] as number[])
     for (const row of tableData) {
-      row[timeColIndex] = new Date(row[timeColIndex]).valueOf()
+      if (timeColIndex >= 0) {
+        row[timeColIndex] = new Date(row[timeColIndex]).valueOf()
+      }
+      if (numberColumns.length) {
+        numberColumns.forEach(colIndex => {
+          if (row[colIndex] !== '') {
+            row[colIndex] = parseFloat(row[colIndex] as string)
+          }
+        })
+      }
     }
 
     const data: Array<Array<string | number>> = [headerRow, ...tableData]